refactor(forges): import settings store directly to avoid barrel import cycle
